Validacoes

+++ src/main/java/br/com/wesleyeduardo/CrudExample/controller/TopicosController.java
@@ -15,6 +15,7 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/topicos")
@@ -55,11 +56,15 @@ public class TopicosController {
 
 
     @GetMapping("/{id}")
-    public DetalhesTopicoDTO detalhar(@PathVariable Long id){
+    public ResponseEntity<DetalhesTopicoDTO> detalhar(@PathVariable Long id){
 
-        Topico topico = topicoRepository.getOne(id);
+        Optional<Topico> topico = topicoRepository.findById(id);
 
-        return new DetalhesTopicoDTO(topico);
+        if(topico.isPresent()){
+            return ResponseEntity.ok(new DetalhesTopicoDTO(topico.get()));
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
 
@@ -67,9 +72,14 @@ public class TopicosController {
     @Transactional
     public ResponseEntity<TopicoDTO> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
 
-       Topico topico =  form.atualizar(id,topicoRepository);
+        Optional<Topico> optional = topicoRepository.findById(id);
 
-        return ResponseEntity.ok(new TopicoDTO(topico));
+        if(optional.isPresent()){
+            Topico topico =  form.atualizar(id,topicoRepository);
+            return ResponseEntity.ok(new TopicoDTO(topico));
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
 
@@ -77,9 +87,16 @@ public class TopicosController {
     @Transactional
     public  ResponseEntity remover(@PathVariable Long id){
 
-        topicoRepository.deleteById(id);
+        Optional<Topico> optional = topicoRepository.findById(id);
 
-        return ResponseEntity.ok().build();
+        if(optional.isPresent()){
+
+            topicoRepository.deleteById(id);
+
+            return ResponseEntity.ok().build();
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
 }
